requirements.txt
tinytuya>=1.8.0
python-dotenv>=0.19.0 